+++ test/simdloop.jl
@@ -115,23 +115,23 @@ function simd_cartesian_range!(indices, crng)
     indices
 end
 
-crng = CartesianRange(2:4, 0:1, 1:1, 3:5)
+crng = CartesianIndices(2:4, 0:1, 1:1, 3:5)
 indices = simd_cartesian_range!(Vector{eltype(crng)}(), crng)
 @test indices == vec(collect(crng))
 
-crng = CartesianRange(-1:1, 1:3)
+crng = CartesianIndices(-1:1, 1:3)
 indices = simd_cartesian_range!(Vector{eltype(crng)}(), crng)
 @test indices == vec(collect(crng))
 
-crng = CartesianRange(-1:-1, 1:3)
+crng = CartesianIndices(-1:-1, 1:3)
 indices = simd_cartesian_range!(Vector{eltype(crng)}(), crng)
 @test indices == vec(collect(crng))
 
-crng = CartesianRange(2:4)
+crng = CartesianIndices(2:4)
 indices = simd_cartesian_range!(Vector{eltype(crng)}(), crng)
 @test indices == collect(crng)
 
-crng = CartesianRange()
+crng = CartesianIndices()
 indices = simd_cartesian_range!(Vector{eltype(crng)}(), crng)
 @test indices == vec(collect(crng))